requirements.txt
Flask==2.3.3
google-cloud-speech==2.23.0
google-cloud-vision==3.7.4
openai==1.41.0
